requirements.txt
pytest>=7.2.1
selenium>=4.8.2
webdriver-manager>=3.8.5
Faker~=17.0.0
requests~=2.28.2
allure-pytest~=2.12.0